refactor(playground): apply React best practices to trace intelligence
Separate complex view concerns into focused components, remove render-time state updates from Sankey motion, and cover reduced-motion and interpolation behavior so the reviewed UI remains maintainable without changing its behavior.\n\nCo-Authored-By: Mastra Code (openai/gpt-5.6-sol) <noreply@mastra.ai>
