+++ packages/playground-ui/src/ds/components/SankeyChart/use-sankey-geometry-transition.ts
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useState, useSyncExternalStore } from 'react';
 
 import type { FixedSankeyGeometry, FixedSankeyLinkGeometry, FixedSankeyNodeGeometry } from './sankey-chart-utils';
 
@@ -7,13 +7,7 @@ const TRANSITION_DURATION_MS = 850;
 type GeometryTransition = {
   from: FixedSankeyGeometry;
   to: FixedSankeyGeometry;
-  progress: number;
-};
-
-type GeometryMotionState = {
-  key?: string;
-  geometry?: FixedSankeyGeometry;
-  transition?: GeometryTransition;
+  startedAt: number;
 };
 
 function mix(from: number, to: number, progress: number) {
@@ -84,6 +78,74 @@ function prefersReducedMotion() {
   );
 }
 
+class SankeyGeometryMotionStore {
+  private animationFrame = 0;
+  private key?: string;
+  private listeners = new Set<() => void>();
+  private snapshot?: FixedSankeyGeometry;
+  private target?: FixedSankeyGeometry;
+  private transition?: GeometryTransition;
+
+  constructor(key?: string, geometry?: FixedSankeyGeometry) {
+    this.key = key;
+    this.snapshot = geometry;
+    this.target = geometry;
+  }
+
+  getSnapshot = () => this.snapshot;
+
+  subscribe = (listener: () => void) => {
+    this.listeners.add(listener);
+    return () => this.listeners.delete(listener);
+  };
+
+  transitionTo(key: string | undefined, geometry: FixedSankeyGeometry | undefined) {
+    if (key === this.key && geometry === this.target) return;
+
+    const keyChanged = key !== this.key;
+    cancelAnimationFrame(this.animationFrame);
+    this.animationFrame = 0;
+    this.key = key;
+    this.target = geometry;
+
+    if (!keyChanged || !geometry || !this.snapshot || prefersReducedMotion()) {
+      this.transition = undefined;
+      this.publish(geometry);
+      return;
+    }
+
+    this.transition = { from: this.snapshot, to: geometry, startedAt: performance.now() };
+    this.animationFrame = requestAnimationFrame(this.animate);
+  }
+
+  dispose() {
+    cancelAnimationFrame(this.animationFrame);
+    this.listeners.clear();
+  }
+
+  private animate = (now: number) => {
+    const transition = this.transition;
+    if (!transition) return;
+
+    const elapsed = Math.min((now - transition.startedAt) / TRANSITION_DURATION_MS, 1);
+    const progress = 1 - (1 - elapsed) ** 3;
+    this.publish(interpolateSankeyGeometry(transition.from, transition.to, progress));
+
+    if (elapsed < 1) {
+      this.animationFrame = requestAnimationFrame(this.animate);
+      return;
+    }
+
+    this.transition = undefined;
+    this.animationFrame = 0;
+  };
+
+  private publish(geometry: FixedSankeyGeometry | undefined) {
+    this.snapshot = geometry;
+    for (const listener of this.listeners) listener();
+  }
+}
+
 export function useSankeyGeometryTransition({
   geometry,
   transitionKey,
@@ -91,45 +153,14 @@ export function useSankeyGeometryTransition({
   geometry?: FixedSankeyGeometry;
   transitionKey?: string;
 }) {
-  const [motion, setMotion] = useState<GeometryMotionState>(() => ({ key: transitionKey, geometry }));
-
-  if (transitionKey !== motion.key) {
-    setMotion({
-      key: transitionKey,
-      geometry,
-      transition:
-        geometry && motion.geometry && !prefersReducedMotion()
-          ? { from: motion.geometry, to: geometry, progress: 0 }
-          : undefined,
-    });
-  } else if (geometry !== motion.geometry && !motion.transition) {
-    setMotion({ ...motion, geometry });
-  }
+  const [store] = useState(() => new SankeyGeometryMotionStore(transitionKey, geometry));
+  const animatedGeometry = useSyncExternalStore(store.subscribe, store.getSnapshot, store.getSnapshot);
 
   useEffect(() => {
-    const from = motion.transition?.from;
-    const to = motion.transition?.to;
-    if (!from || !to) return;
+    store.transitionTo(transitionKey, geometry);
+  }, [geometry, store, transitionKey]);
 
-    const startedAt = performance.now();
-    let animationFrame = 0;
-    const animate = (now: number) => {
-      const elapsed = Math.min((now - startedAt) / TRANSITION_DURATION_MS, 1);
-      const progress = 1 - (1 - elapsed) ** 3;
-      setMotion(current => {
-        if (current.transition?.from !== from || current.transition.to !== to) return current;
-        if (elapsed === 1) return { key: current.key, geometry: to };
-        return { ...current, transition: { from, to, progress } };
-      });
-      if (elapsed < 1) animationFrame = requestAnimationFrame(animate);
-    };
-    animationFrame = requestAnimationFrame(animate);
+  useEffect(() => () => store.dispose(), [store]);
 
-    return () => cancelAnimationFrame(animationFrame);
-  }, [motion.transition?.from, motion.transition?.to]);
-
-  const transition = motion.transition;
-  return transition
-    ? interpolateSankeyGeometry(transition.from, transition.to, transition.progress)
-    : (geometry ?? motion.geometry);
+  return animatedGeometry;
 }
