+++ packages/playground/src/ee/signals/sankey-signals.tsx
@@ -1,23 +1,12 @@
-import type { DragStart, DragUpdate, DraggableStateSnapshot, DropResult, DroppableProvided } from '@hello-pangea/dnd';
-import { DragDropContext, Draggable, Droppable, useMouseSensor, useTouchSensor } from '@hello-pangea/dnd';
-import { Button } from '@mastra/playground-ui/components/Button';
-import { Card, CardContent } from '@mastra/playground-ui/components/Card';
-import { nodeColor, Sankey, SankeyChart } from '@mastra/playground-ui/components/SankeyChart';
-import type {
-  SankeyChartColumn,
-  SankeyChartNodeSelection,
-  SankeyChartRecord,
-} from '@mastra/playground-ui/components/SankeyChart';
-import { Tab, TabList, Tabs } from '@mastra/playground-ui/components/Tabs';
-import { Tooltip, TooltipContent, TooltipTrigger } from '@mastra/playground-ui/components/Tooltip';
-import { Txt } from '@mastra/playground-ui/components/Txt';
-import { getSignalHue, SignalsOverviewPage as SignalsEmptyState } from '@mastra/playground-ui/ee/signals';
-import { Icon } from '@mastra/playground-ui/icons/Icon';
+import type { SankeyChartNodeSelection } from '@mastra/playground-ui/components/SankeyChart';
+import { TabList, Tabs } from '@mastra/playground-ui/components/Tabs';
+import { SignalsOverviewPage as SignalsEmptyState } from '@mastra/playground-ui/ee/signals';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { ArrowLeftRight, ChartNoAxesGantt, GripVertical, Info, Waypoints, X } from 'lucide-react';
+import { ArrowLeftRight, ChartNoAxesGantt, Waypoints } from 'lucide-react';
 import { useMemo, useState } from 'react';
 
 import { fetchThemeFlow, fetchThemePaths, fetchThemeSnapshots } from './entity-learning-api';
+import { FlowCard } from './flow-card';
 import { useEntityLearningProgress } from './hooks/use-entity-learning-progress';
 import { useSnapshotPlayback } from './hooks/use-snapshot-playback';
 import { useThemeFlows } from './hooks/use-theme-flows';
@@ -26,14 +15,10 @@ import { useThemeSnapshots } from './hooks/use-theme-snapshots';
 import { NoiseDetailPanel } from './noise-detail-panel';
 import {
   buildSignalGraphSummary,
-  getSignalRecordNodeId,
-  getSignalRecordNodeLabel,
-  getSignalRecordNodeValue,
   selectFlowSnapshotIds,
   snapshotSummaryLabel,
   stabilizeThemeFlow,
 } from './sankey-signals-data';
-import { formatSignalName, getSignalDescription, SIGNAL_DESCRIPTIONS } from './signal-formatting';
 import { SignalsErrorState } from './signals-error-state';
 import { SignalsFrameLoadingSkeleton, SignalsLoadingSkeleton } from './signals-loading-skeleton';
 import { SnapshotTimeline } from './snapshot-timeline';
@@ -41,8 +26,12 @@ import { ThemeCompare } from './theme-compare';
 import { ThemeDetailPanel } from './theme-detail-panel';
 import { buildDrilledThemeFlow, findNoiseSelection, findThemeSelection } from './theme-drilldown-data';
 import type { ThemeSelection } from './theme-drilldown-data';
+import { ThemeFilterBanner } from './theme-filter-banner';
 import { ThemeLifelines } from './theme-lifelines';
-import type { ThemeFlowResponse, TraceSignalName } from './types';
+import { TraceIntelligenceExplainer } from './trace-intelligence-explainer';
+import type { TraceSignalName } from './types';
+import { ViewModeTab } from './view-mode-tab';
+import type { SignalsViewMode } from './view-mode-tab';
 import { Link } from '@/lib/link';
 
 export interface SankeySignalsProps {
@@ -58,342 +47,12 @@ export interface SankeySignalsProps {
 
 const DRILL_IN_TRACE_LIMIT = 2000;
 
-const DRAG_SENSORS = [useMouseSensor, useTouchSensor];
-
-type SignalsViewMode = 'flow' | 'compare' | 'lifelines';
-
 /** One-line answer to "what am I looking at?" for each view, shown under the tabs. */
 const VIEW_DESCRIPTIONS: Record<SignalsViewMode, string> = {
   flow: "How this agent's traces distribute across goal, sentiment, behavior, and outcome themes at this point in time.",
   compare: 'Which themes grew, shrank, appeared, or disappeared between two points in time.',
   lifelines: "Each theme's share of traces across the whole selected range.",
 };
-
-const EXPLAINER_SIGNAL_ORDER: TraceSignalName[] = ['goal', 'sentiment', 'behavior', 'outcome'];
-
-/** Info tooltip for first-time viewers: signals → themes → snapshots. */
-function TraceIntelligenceExplainer() {
-  return (
-    <Tooltip>
-      <TooltipTrigger
-        aria-label="What is trace intelligence?"
-        className="text-neutral3 hover:text-neutral6 flex cursor-help items-center transition-colors"
-        type="button"
-      >
-        <Icon size="sm">
-          <Info />
-        </Icon>
-      </TooltipTrigger>
-      <TooltipContent className="max-w-sm space-y-3 p-4 text-xs">
-        <p className="text-neutral5">
-          Every trace from this agent is analyzed for four signals, and traces with similar signals are clustered into
-          named themes.
-        </p>
-        <ul className="space-y-1.5">
-          {EXPLAINER_SIGNAL_ORDER.map(signalName => (
-            <li key={signalName} className="text-neutral4">
-              <span
-                className="font-mono text-[10px] font-semibold tracking-widest uppercase"
-                style={{ color: nodeColor(getSignalHue(signalName)) }}
-              >
-                {signalName}
-              </span>{' '}
-              — {SIGNAL_DESCRIPTIONS[signalName]}
-            </li>
-          ))}
-        </ul>
-        <p className="text-neutral4">
-          Snapshots capture the themes at points in time, so the views show how they appear, grow, and fade.
-        </p>
-      </TooltipContent>
-    </Tooltip>
-  );
-}
-
-function ViewModeTab({ value, icon, label }: { value: SignalsViewMode; icon: React.ReactNode; label: string }) {
-  return (
-    <Tab value={value} className="px-3 py-2">
-      <Icon size="sm">{icon}</Icon>
-      <Txt variant="ui-sm" className="text-inherit">
-        {label}
-      </Txt>
-    </Tab>
-  );
-}
-
-/**
- * Active drill-in state banner: a dismissible filter chip in the theme's
- * signal hue plus a plain-language description of the filtered subset, so the
- * chart below clearly reads as "traces flowing through this theme" rather
- * than a full snapshot.
- */
-function ThemeFilterBanner({
-  selection,
-  filteredTraceCount,
-  totalTraceCount,
-  onViewDetails,
-  onClear,
-}: {
-  selection: ThemeSelection;
-  filteredTraceCount?: number;
-  totalTraceCount: number;
-  onViewDetails: () => void;
-  onClear: () => void;
-}) {
-  const color = nodeColor(getSignalHue(selection.signalName));
-
-  return (
-    <section
-      aria-label="Active theme drill-in"
-      className="flex flex-wrap items-center gap-x-4 gap-y-2 rounded-lg border px-3 py-2"
-      style={{
-        borderColor: `color-mix(in srgb, ${color} 35%, transparent)`,
-        backgroundColor: `color-mix(in srgb, ${color} 8%, transparent)`,
-      }}
-    >
-      <button
-        aria-label="Clear theme filter"
-        className="border-border1 bg-surface2 text-neutral6 hover:bg-surface4 flex items-center gap-1.5 rounded-full border py-1 pr-2 pl-2.5 text-xs font-medium transition-colors"
-        onClick={onClear}
-        type="button"
-      >
-        <span aria-hidden="true" className="size-2 rounded-[2px]" style={{ backgroundColor: color }} />
-        {formatSignalName(selection.signalName)} · {selection.label}
-        <X aria-hidden="true" className="size-3.5" />
-      </button>
-      <span className="text-neutral4 text-xs">
-        {filteredTraceCount === undefined
-          ? 'Loading theme traces…'
-          : `Showing the ${filteredTraceCount} of ${totalTraceCount} traces that flow through this theme`}
-      </span>
-      <Button
-        aria-label={`View theme details for ${selection.label}`}
-        onClick={onViewDetails}
-        size="sm"
-        type="button"
-        variant="ghost"
-      >
-        Details →
-      </Button>
-    </section>
-  );
-}
-
-function FlowCard({
-  columns,
-  records,
-  stages,
-  height,
-  onNodeClick,
-  isNodeClickable,
-  drillInDisabledReason,
-  onOrderChange,
-  signalOrder,
-  reorderDisabled,
-}: {
-  columns: SankeyChartColumn[];
-  records: SankeyChartRecord[];
-  stages: ThemeFlowResponse['stages'];
-  height?: number;
-  onNodeClick?: (selection: SankeyChartNodeSelection) => void;
-  isNodeClickable?: (selection: SankeyChartNodeSelection) => boolean;
-  drillInDisabledReason?: string;
-  onOrderChange: (signalNames: TraceSignalName[]) => void;
-  signalOrder: TraceSignalName[];
-  reorderDisabled: boolean;
-}) {
-  const chartColumns = columns.map(column => ({ ...column, label: column.label.toUpperCase() }));
-  // Keep the dropped header order visible while the matching chart data loads.
-  const linkedColumnIds = new Set(columns.map(column => column.id));
-  const orderedLinkedSignals = signalOrder.filter(signalName => linkedColumnIds.has(signalName));
-  const orderedSignalSet = new Set(orderedLinkedSignals);
-  const headerSignalNames = [
-    ...orderedLinkedSignals,
-    ...stages
-      .map(stage => stage.signalName)
-      .filter(signalName => linkedColumnIds.has(signalName) && !orderedSignalSet.has(signalName)),
-  ];
-  const [dragProjection, setDragProjection] = useState<{ sourceIndex: number; destinationIndex: number }>();
-
-  const handleDragStart = (start: DragStart) => {
-    setDragProjection({ sourceIndex: start.source.index, destinationIndex: start.source.index });
-  };
-
-  const handleDragUpdate = (update: DragUpdate) => {
-    setDragProjection({
-      sourceIndex: update.source.index,
-      destinationIndex: update.destination?.index ?? update.source.index,
-    });
-  };
-
-  const handleDragEnd = (result: DropResult) => {
-    setDragProjection(undefined);
-    const destinationIndex = result.destination?.index;
-    if (destinationIndex === undefined || destinationIndex === result.source.index) return;
-
-    const reordered = [...headerSignalNames];
-    const [movedSignalName] = reordered.splice(result.source.index, 1);
-    if (!movedSignalName) return;
-    reordered.splice(destinationIndex, 0, movedSignalName);
-    // Signals without a chart column still belong to the perspective; keep
-    // them in the request after the reordered columns.
-    const seen = new Set<TraceSignalName>(reordered);
-    onOrderChange([...reordered, ...stages.map(stage => stage.signalName).filter(name => !seen.has(name))]);
-  };
-
-  return (
-    <Card
-      aria-label="Trace signal theme flow"
-      as="section"
-      className="relative min-w-0"
-      elevation="elevated"
-      title={drillInDisabledReason}
-    >
-      <span
-        aria-hidden="true"
-        className="bg-surface2 text-neutral3 absolute top-0 left-5 -translate-y-1/2 px-2 font-mono text-[10px] tracking-[0.18em]"
-      >
-        SIGNALS
-      </span>
-      <CardContent className="px-0 pt-4 pb-2 sm:pt-5 sm:pb-3">
-        <div aria-label="Signals" role="group">
-          <DragDropContext
-            enableDefaultSensors={false}
-            sensors={DRAG_SENSORS}
-            onDragEnd={handleDragEnd}
-            onDragStart={handleDragStart}
-            onDragUpdate={handleDragUpdate}
-          >
-            <Droppable direction="horizontal" droppableId="signal-column-headers">
-              {(provided: DroppableProvided) => (
-                <div
-                  {...provided.droppableProps}
-                  ref={provided.innerRef}
-                  aria-label="Trace signal column headers"
-                  className="flex items-center gap-1 px-8 pb-1"
-                  role="group"
-                >
-                  {headerSignalNames.map((signalName, index) => {
-                    const label = formatSignalName(signalName);
-                    let projectedIndex = index;
-                    if (dragProjection) {
-                      const { sourceIndex, destinationIndex } = dragProjection;
-                      if (index === sourceIndex) projectedIndex = destinationIndex;
-                      else if (sourceIndex < destinationIndex && index > sourceIndex && index <= destinationIndex)
-                        projectedIndex = index - 1;
-                      else if (destinationIndex < sourceIndex && index >= destinationIndex && index < sourceIndex)
-                        projectedIndex = index + 1;
-                    }
-                    const offsetPercent =
-                      headerSignalNames.length > 1 ? (projectedIndex / (headerSignalNames.length - 1) - 0.5) * 100 : 0;
-                    const headerAnchor =
-                      projectedIndex === 0
-                        ? 'start'
-                        : projectedIndex === headerSignalNames.length - 1
-                          ? 'end'
-                          : 'middle';
-                    const contentOffsetClass =
-                      headerAnchor === 'start' ? 'translate-x-1/2' : headerAnchor === 'end' ? '-translate-x-1/2' : '';
-                    return (
-                      <Draggable
-                        key={signalName}
-                        draggableId={signalName}
-                        index={index}
-                        isDragDisabled={reorderDisabled}
-                      >
-                        {(dragProvided, dragSnapshot: DraggableStateSnapshot) => (
-                          <div
-                            ref={dragProvided.innerRef}
-                            {...dragProvided.draggableProps}
-                            className="flex min-w-0 flex-1 basis-0 items-center justify-center py-1"
-                            data-dragging={dragSnapshot.isDragging}
-                            style={dragProvided.draggableProps.style}
-                          >
-                            <div
-                              className="flex w-full justify-center"
-                              data-testid="signal-column-header-alignment"
-                              style={{ translate: `${offsetPercent}%` }}
-                            >
-                              <div
-                                className={`relative inline-flex items-center justify-center rounded-md border border-transparent px-1 py-0.5 motion-safe:transition-[background-color,border-color,box-shadow,scale] motion-safe:duration-150 ${contentOffsetClass} ${
-                                  dragSnapshot.isDragging ? 'border-border2 bg-surface4 scale-[1.03] shadow-lg' : ''
-                                }`}
-                                data-header-anchor={headerAnchor}
-                                data-testid="signal-column-header-content"
-                              >
-                                <Tooltip>
-                                  <TooltipTrigger
-                                    className="cursor-default font-mono text-xs font-semibold tracking-wider"
-                                    data-testid="signal-column-header"
-                                    style={{ color: nodeColor(getSignalHue(signalName)) }}
-                                  >
-                                    {label.toUpperCase()}
-                                  </TooltipTrigger>
-                                  <TooltipContent>{getSignalDescription(signalName)}</TooltipContent>
-                                </Tooltip>
-                                <div
-                                  {...dragProvided.dragHandleProps}
-                                  aria-disabled={reorderDisabled}
-                                  aria-label={`Reorder ${label}`}
-                                  className="text-neutral3 hover:text-neutral5 absolute top-1/2 ml-0.5 -translate-y-1/2 cursor-grab rounded-sm p-1 active:cursor-grabbing aria-disabled:cursor-wait aria-disabled:opacity-50"
-                                  style={{ left: '100%' }}
-                                  title={`Drag to reorder the ${label} column`}
-                                >
-                                  <GripVertical aria-hidden="true" className="size-3.5" />
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-                        )}
-                      </Draggable>
-                    );
-                  })}
-                  {provided.placeholder}
-                </div>
-              )}
-            </Droppable>
-          </DragDropContext>
-        </div>
-        <div
-          aria-label="Themes"
-          className="text-neutral3 flex items-center gap-2 py-1 font-mono text-[10px] tracking-[0.18em]"
-          role="separator"
-        >
-          <span aria-hidden="true" className="bg-border1 h-px w-5" />
-          THEMES
-          <span aria-hidden="true" className="bg-border1 h-px flex-1" />
-        </div>
-        <div aria-busy={reorderDisabled} data-testid="sankey-order-transition">
-          <Sankey
-            data={records}
-            columns={chartColumns}
-            columnOrder={chartColumns.map(column => column.id)}
-            getColumnHue={column => getSignalHue(column.id)}
-            getRecordNodeId={getSignalRecordNodeId}
-            getRecordNodeLabel={getSignalRecordNodeLabel}
-            getRecordNodeValue={getSignalRecordNodeValue}
-            getRecordWeight={record => Number(record.traceCount)}
-            getRecordLayoutWeight={record => Number(record.layoutTraceCount)}
-          >
-            <SankeyChart
-              height={height ?? 'clamp(340px, 42vw, 460px)'}
-              margin={{ top: 40, right: 32, bottom: 24, left: 32 }}
-              onNodeClick={onNodeClick}
-              isNodeClickable={isNodeClickable}
-              hideColumnLabels
-              geometryTransitionKey={chartColumns.map(column => column.id).join(':')}
-            />
-          </Sankey>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-/** Right-aligned control row shown when the view tabs are unavailable. */
-function DateRangeRow({ children }: { children: React.ReactNode }) {
-  return <div className="flex justify-end px-4 pt-4 lg:px-6 lg:pt-6">{children}</div>;
-}
 
 export function SankeySignals({
   entityId,
@@ -537,7 +196,7 @@ export function SankeySignals({
   if (snapshotsQuery.isPending) {
     return (
       <>
-        {dateRangePicker && <DateRangeRow>{dateRangePicker}</DateRangeRow>}
+        {dateRangePicker && <div className="flex justify-end px-4 pt-4 lg:px-6 lg:pt-6">{dateRangePicker}</div>}
         <SignalsLoadingSkeleton />
       </>
     );
@@ -546,7 +205,7 @@ export function SankeySignals({
   if (snapshotsQuery.isError || hasFlowError || hasActivePathsError) {
     return (
       <>
-        {dateRangePicker && <DateRangeRow>{dateRangePicker}</DateRangeRow>}
+        {dateRangePicker && <div className="flex justify-end px-4 pt-4 lg:px-6 lg:pt-6">{dateRangePicker}</div>}
         <SignalsErrorState
           message="Unable to load trace signal flow."
           onRetry={() => {
@@ -564,7 +223,7 @@ export function SankeySignals({
   if (!snapshot) {
     return (
       <>
-        {dateRangePicker && <DateRangeRow>{dateRangePicker}</DateRangeRow>}
+        {dateRangePicker && <div className="flex justify-end px-4 pt-4 lg:px-6 lg:pt-6">{dateRangePicker}</div>}
         <SignalsEmptyState LinkComponent={Link} progress={progressQuery.data} isRangeEmpty />
       </>
     );
@@ -591,7 +250,7 @@ export function SankeySignals({
   if (!currentFlow || !flow || !graphSummary || populatedStageCount < 2) {
     return (
       <>
-        {dateRangePicker && <DateRangeRow>{dateRangePicker}</DateRangeRow>}
+        {dateRangePicker && <div className="flex justify-end px-4 pt-4 lg:px-6 lg:pt-6">{dateRangePicker}</div>}
         <SignalsEmptyState LinkComponent={Link} progress={progressQuery.data} />
       </>
     );
