+++ packages/playground/src/ee/signals/theme-compare.tsx
@@ -1,166 +1,14 @@
-import { nodeColor } from '@mastra/playground-ui/components/SankeyChart';
-import { Tooltip, TooltipContent, TooltipTrigger } from '@mastra/playground-ui/components/Tooltip';
-import { getSignalHue } from '@mastra/playground-ui/ee/signals';
 import { useState } from 'react';
 
 import { useThemeFlows } from './hooks/use-theme-flows';
 import { snapshotSummaryLabel } from './sankey-signals-data';
-import { formatSignalName, SIGNAL_DESCRIPTIONS } from './signal-formatting';
+import { SignalDeltaColumn } from './signal-delta-column';
 import { SignalsFrameLoadingSkeleton } from './signals-loading-skeleton';
 import { TimelineTrack } from './snapshot-timeline';
 import type { TimelineMarkerKind } from './snapshot-timeline';
 import { timelineTickPositions } from './snapshot-timeline-data';
-import { computeThemeShareDeltas, themeShareSeries } from './theme-compare-data';
 import type { ThemeSelection } from './theme-drilldown-data';
-import type { ThemeFlowResponse, ThemeSnapshot, TraceSignalName } from './types';
-
-const SPARKLINE_WIDTH = 100;
-const SPARKLINE_HEIGHT = 20;
-
-function percent(share: number) {
-  return `${Math.round(share * 100)}%`;
-}
-
-function deltaLabel(delta: number) {
-  const points = Math.round(delta * 100);
-  return `${points >= 0 ? '+' : ''}${points}%`;
-}
-
-function Sparkline({
-  series,
-  positions,
-  markerIndexes,
-}: {
-  series: Array<number | undefined>;
-  positions: number[];
-  markerIndexes: number[];
-}) {
-  const loaded = series.flatMap((share, index) => (share === undefined ? [] : [{ share, index }]));
-  if (loaded.length < 2) return null;
-
-  const maxShare = Math.max(...loaded.map(point => point.share), 0.01);
-  const pointFor = (point: { share: number; index: number }) => ({
-    x: (positions[point.index]! / 100) * SPARKLINE_WIDTH,
-    y: SPARKLINE_HEIGHT - 2 - (point.share / maxShare) * (SPARKLINE_HEIGHT - 4),
-  });
-  const polyline = loaded.map(point => `${pointFor(point).x.toFixed(1)},${pointFor(point).y.toFixed(1)}`).join(' ');
-  const markers = markerIndexes.flatMap(markerIndex => {
-    const point = loaded.find(candidate => candidate.index === markerIndex);
-    return point ? [point] : [];
-  });
-
-  // The svg stretches horizontally (preserveAspectRatio none), which would
-  // flatten circles — so the compare dots render as HTML overlays that stay
-  // round. The viewBox height matches the h-5 track, so y coordinates are pixels.
-  return (
-    <div aria-hidden="true" className="relative mt-1.5 h-5 w-full">
-      <svg
-        className="absolute inset-0 size-full"
-        preserveAspectRatio="none"
-        viewBox={`0 0 ${SPARKLINE_WIDTH} ${SPARKLINE_HEIGHT}`}
-      >
-        <polyline
-          className="stroke-neutral3 fill-none"
-          points={polyline}
-          strokeWidth="1"
-          vectorEffect="non-scaling-stroke"
-        />
-      </svg>
-      {markers.map(point => (
-        <span
-          key={point.index}
-          className="absolute size-2 -translate-x-1/2 -translate-y-1/2 rounded-full bg-green-400"
-          style={{ left: `${positions[point.index]}%`, top: `${pointFor(point).y}px` }}
-        />
-      ))}
-    </div>
-  );
-}
-
-function SignalDeltaColumn({
-  signalName,
-  fromFlow,
-  toFlow,
-  flows,
-  positions,
-  fromIndex,
-  toIndex,
-  onThemeSelect,
-}: {
-  signalName: TraceSignalName;
-  fromFlow: ThemeFlowResponse;
-  toFlow: ThemeFlowResponse;
-  flows: Array<ThemeFlowResponse | undefined>;
-  positions: number[];
-  fromIndex: number;
-  toIndex: number;
-  onThemeSelect: (selection: ThemeSelection, snapshotIndex: number) => void;
-}) {
-  const deltas = computeThemeShareDeltas(fromFlow, toFlow, signalName);
-  // Details open at the compared snapshot where the theme still exists.
-  const detailIndexFor = (delta: { toShare: number }) => (delta.toShare > 0 ? toIndex : fromIndex);
-
-  return (
-    <section aria-label={`${formatSignalName(signalName)} changes`} className="min-w-0">
-      <h3
-        className="font-mono text-xs font-semibold tracking-widest uppercase"
-        style={{ color: nodeColor(getSignalHue(signalName)) }}
-      >
-        <Tooltip>
-          <TooltipTrigger className="cursor-default uppercase">{signalName}</TooltipTrigger>
-          <TooltipContent>{SIGNAL_DESCRIPTIONS[signalName]}</TooltipContent>
-        </Tooltip>
-      </h3>
-      <ul className="mt-2 space-y-1.5">
-        {deltas.length === 0 ? <li className="text-neutral3 text-xs">No themes in either snapshot.</li> : null}
-        {deltas.map(delta => {
-          const themeId = delta.themeId;
-          const card = (
-            <>
-              <div className="flex items-baseline justify-between gap-2">
-                <span className="text-neutral6 truncate text-xs font-medium" title={delta.label}>
-                  {delta.label}
-                </span>
-                <span className="text-neutral6 shrink-0 font-mono text-xs font-semibold tabular-nums">
-                  {deltaLabel(delta.delta)}
-                </span>
-              </div>
-              <p className="text-neutral3 font-mono text-[11px] tabular-nums">
-                {percent(delta.fromShare)} → {percent(delta.toShare)}
-              </p>
-              <Sparkline
-                series={themeShareSeries(flows, signalName, delta.label)}
-                positions={positions}
-                markerIndexes={[fromIndex, toIndex]}
-              />
-            </>
-          );
-          return (
-            <li
-              key={delta.label}
-              className={`border-border1 rounded-lg border ${
-                delta.delta > 0 ? 'bg-green-500/5' : delta.delta < 0 ? 'bg-red-500/5' : 'bg-surface3'
-              }`}
-            >
-              {themeId === undefined ? (
-                <div className="px-2.5 py-2">{card}</div>
-              ) : (
-                <button
-                  aria-label={`View theme details for ${delta.label}`}
-                  className="hover:border-border2 block w-full cursor-pointer rounded-lg px-2.5 py-2 text-left transition-colors hover:bg-white/[0.03]"
-                  onClick={() => onThemeSelect({ signalName, themeId, label: delta.label }, detailIndexFor(delta))}
-                  type="button"
-                >
-                  {card}
-                </button>
-              )}
-            </li>
-          );
-        })}
-      </ul>
-    </section>
-  );
-}
+import type { ThemeSnapshot, TraceSignalName } from './types';
 
 /**
  * Compare mode: two interchangeable points on the shared time axis show how
@@ -223,9 +71,14 @@ export function ThemeCompare({
       setGrabbedPoint(pointAtIndex);
       return;
     }
-    const distanceTo = (point: 0 | 1) => Math.abs(positions[index]! - positions[points[point]]!);
-    const nearest = distanceTo(0) <= distanceTo(1) ? 0 : 1;
-    setPointIndexes(nearest === 0 ? [index, points[1]] : [points[0], index]);
+    const selectedPosition = positions[index];
+    const firstPointPosition = positions[points[0]];
+    const secondPointPosition = positions[points[1]];
+    if (selectedPosition === undefined || firstPointPosition === undefined || secondPointPosition === undefined) return;
+
+    const distanceToFirst = Math.abs(selectedPosition - firstPointPosition);
+    const distanceToSecond = Math.abs(selectedPosition - secondPointPosition);
+    setPointIndexes(distanceToFirst <= distanceToSecond ? [index, points[1]] : [points[0], index]);
   };
 
   const markers = new Map<number, TimelineMarkerKind>([
